consolidate to single file
